DAC6-3052: Hooked up page not found functionality to error handler
The updated "Page not found" page was set up to only display on a specific page-not-found url. This is incorrect, it should display on any URL which isn't valid.

Removed the specific route for this page and hooked it into the error handler so it will display for any invalid url.